feat: poc test
