+++ packages/datasource-customizer/src/decorators/computed/types.ts
@@ -1,4 +1,10 @@
-import { ColumnType } from '@forestadmin/datasource-toolkit';
+import {
+  ColumnType,
+  ManyToManySchema,
+  ManyToOneSchema,
+  OneToManySchema,
+  OneToOneSchema,
+} from '@forestadmin/datasource-toolkit';
 
 import CollectionCustomizationContext from '../../context/collection-context';
 import { TCollectionName, TFieldName, TRow, TSchema } from '../../templates';
@@ -24,3 +30,10 @@ export interface DeprecatedComputedDefinition<
 > extends Omit<ComputedDefinition<S, N>, 'columnType'> {
   readonly columnType: 'Timeonly';
 }
+
+type PartialBy<T, K extends Extract<keyof T, string>> = Omit<T, K> & Partial<Pick<T, K>>;
+
+export type RelationDefinition =
+  | PartialBy<ManyToOneSchema, 'foreignKeyTarget'>
+  | PartialBy<OneToManySchema | OneToOneSchema, 'originKeyTarget'>
+  | PartialBy<ManyToManySchema, 'foreignKeyTarget' | 'originKeyTarget'>;
